+++ src/components/AutocompleteDropdown.tsx
@@ -306,7 +306,7 @@ function AutocompleteDropdown({
       </div>
 
       {(isDirty || displayValidationMessage) && !!validationMessage && (
-        <div role="alert">
+        <div role="alert" className="has-margin-top-8">
           <div className="has-text-danger ob-error__text cypress-validation-message">
             {validationMessage}
           </div>