+++ src/form-elements/FormElementFiles.tsx
@@ -229,7 +229,7 @@ function FormElementFiles({
         </div>
 
         {(isDirty || displayValidationMessage) && !!validationMessage && (
-          <div role="alert">
+          <div role="alert" className="has-margin-top-8">
             <div className="has-text-danger ob-error__text cypress-validation-message">
               {validationMessage}
             </div>