+++ src/form-elements/FormElementNumber.tsx
@@ -101,7 +101,7 @@ function FormElementNumber({
         )}
 
         {(isDirty || displayValidationMessage) && !!validationMessage && (
-          <div role="alert">
+          <div role="alert" className="has-margin-top-8">
             <div className="has-text-danger ob-error__text cypress-validation-message">
               {validationMessage}
             </div>