+++ src/form-elements/FormElementTelephone.tsx
@@ -68,7 +68,7 @@ function FormElementTelephone({
         </div>
 
         {(isDirty || displayValidationMessage) && !!validationMessage && (
-          <div role="alert">
+          <div role="alert" className="has-margin-top-8">
             <div className="has-text-danger ob-error__telephone cypress-validation-message">
               {validationMessage}
             </div>